Fixing merge issue from PR#774
